fix add collabs button
