+++ client/mobile/components/documents/add_collaborators_container.js
@@ -1,4 +1,5 @@
 import { connect } from 'react-redux';
+import { addCollaborator } from '../../../shared/actions/document_actions';
 import AddCollaborators from './add_collaborators';
 
 const mapStateToProps = (state) => ({
@@ -7,8 +8,9 @@ const mapStateToProps = (state) => ({
 
 const mapDispatchToProps = dispatch => ({
   // addCollaborator: () => dispatch(addCollaborators())
-  addCollaborator: newCollaborator => console.log('add colllabs yo'),
-  removeCollaborator: () => console.log('remove collabs yo')
+  removeCollaborator: () => console.log('remove collabs yo'),
+  addCollaborator: (document, user) => dispatch(addCollaborator(document, user))
+
 });
 
 const AddCollaboratorsContainer = connect(
