+++ client/mobile/components/documents/document_frame.js
@@ -1,50 +1,8 @@
 import React from 'react';
-import { View, StatusBar, Button } from 'react-native';
-import { connect } from 'react-redux';
+import { View } from 'react-native';
 import DocumentContainer from './document_container';
-import styles from '../../styles';
-import { selectDocumentAction, toggle, selectColor } from '../../../shared/actions/ui_actions';
-import AddCollaborators from './add_collaborators';
-
-
-// prop is sent to document tools instead of DocumentFrame in the container
-// go fix it
-const DocumentTools = ({ navigation, selectAction }) => (
-  <View>
-    <StatusBar
-      barStyle="dark-content"
-    />
-    <View style={styles.stackNavigation}>
-      <Button
-        title="Type"
-        onPress={() => selectAction('typing')}
-      />
-      <Button
-        title="Draw"
-        onPress={() => selectAction('drawing')}
-      />
-      {/* some action to handle color change */}
-      <Button
-        title="black"
-        onPress={() => selectColor('black')}
-      />
-      {/* some action to handle erasing */}
-      <Button
-        title="eraser"
-        onPress={() => toggle('eraseme')}
-      />
-
-    </View>
-  </View>
-);
-
-const mapProps = dispatch => ({
-  selectAction: (action) => dispatch(selectDocumentAction(action)),
-  toggle: (uiElement) => dispatch(toggle(uiElement)),
-  selectColor: (color) => dispatch(selectColor(color))
-});
-
-const DocumentToolsContainer = connect(null, mapProps)(DocumentTools);
+import DocumentToolsContainer from './document_tools_container';
+// import DocumentTools from './document_tools';
 
 class DocumentFrame extends React.Component {
   static get navigationOptions() {
@@ -54,11 +12,13 @@ class DocumentFrame extends React.Component {
   }
 
   render() {
-    return <DocumentContainer />;
+    const { navigation } = this.props;
+    return (
+
+      <DocumentContainer />
+
+    );
   }
 }
 
 export default DocumentFrame;
-
-// make some new actions, make a reducer for the actions
-// add to container
