+++ client/mobile/navigation.js
@@ -2,6 +2,7 @@ import { createStackNavigator } from 'react-navigation';
 import HomeContainer from './components/home/home_container';
 import Welcome from './components/splash/welcome';
 import Splash from './components/splash/splash';
+import DocumentTools from './components/documents/document_tools';
 import DocumentFrameContainer from './components/documents/document_frame_container';
 import AddCollaboratorsContainer from './components/documents/add_collaborators_container';
 
@@ -17,7 +18,8 @@ export const AppStack = createStackNavigator(
   {
     HomeContainer,
     DocumentFrameContainer,
-    AddCollaboratorsContainer
+    AddCollaboratorsContainer,
+    DocumentTools
   },
   { initialRouteName: 'HomeContainer' }
 );
